requirements.txt
# Core dependencies for Voynich Morphemic Decryption
# Note: For development, use Poetry (see pyproject.toml)
# This file is for compatibility with pip-based workflows

numpy>=1.26.0,<2.0.0
pandas>=2.1.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
scipy>=1.11.0,<2.0.0
xgboost>=2.0.0,<3.0.0
lightgbm>=4.1.0,<5.0.0
fastapi>=0.104.0,<1.0.0
pydantic>=2.4.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
uvicorn[standard]>=0.24.0,<1.0.0
plotly>=5.17.0,<6.0.0
matplotlib>=3.8.0,<4.0.0
seaborn>=0.13.0,<1.0.0
spacy>=3.7.0,<4.0.0
nltk>=3.8.1,<4.0.0
gensim>=4.3.0,<5.0.0
statsmodels>=0.14.0,<1.0.0
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
pyyaml>=6.0.1,<7.0.0
typer[all]>=0.9.0,<1.0.0
rich>=13.7.0,<14.0.0
